amos/django_orchestrator/requirements.txt
django
djangorestframework
django-swagger-render
requests
psycopg2-binary
Image